pyproject.toml
[project]
name = "graphlayout2image"
version = "0.1.0"
description = "Scene Graph Conditioned Image Generation using GNNs and Diffusion Models"
authors = [{ name = "Zeinab Sheikhi", email = "zeinab.sheykhie@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torch-geometric>=2.4.0",
    "pillow>=10.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "wandb>=0.16.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["graphlayout2image"]

[project.scripts]
generate-data = "graphlayout2image.data_generator:main"
